Working on the tests
