+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "env-cmd -f ./config/dev.env  nodemon --watch src --exec babel-node src/index.js",
     "dev": "env-cmd -f ./config/dev.env  nodemon src/index.js",
-    "test": "env-cmd -f ./config/test.env jest --detectOpenHandles --coverage",
-    "travis": "jest --coverage",
+    "test": "env-cmd -f ./config/test.env  jest --forceExit --detectOpenHandles --coverage",
+    "travis": "jest  --forceExit --detectOpenHandles --coverage",
     "lint": "eslint src/"
   },
   "repository": {
