+++ src/tests/routes/api/articles/article.test.js
@@ -1,12 +1,12 @@
 const request = require('supertest');
-const app = require('../../../../app');
-const Article = require('../../../../models/Article');
-const User = require('../../../../models/User');
+import app from '../../../../index';
+import Article from '../../../../models/Article';
+import User from '../../../../models/User';
 const mongoose = require('mongoose');
 
 describe('Tests for validating the create articles feature', () => {
     process.env.API_BASE = '/api';
-    const userApiBase = process.env.API_BASE + '/users';
+    const userApiBase = process.env.API_BASE;
     const articleApiBase = process.env.API_BASE + '/articles';
     let access_token;
 
@@ -21,7 +21,7 @@ describe('Tests for validating the create articles feature', () => {
                 password: 'P@ssw0rd',
                 confirmPassword: 'P@ssw0rd',
             });
-        res = await request(app)
+        const res = await request(app)
             .post(userApiBase + '/login')
             .send({
                 email: 'nserekopaul@gmail.com',
@@ -378,6 +378,6 @@ describe('Tests for validating the create articles feature', () => {
             .delete(articleApiBase + '/comment/javascript/5678w3657w')
             .set('Authorization', `${access_token}`)
             .expect(404);
-        expect(response.body.error).toBe('Comment does not exist');
+        expect(response.body.message).toBe('Resource not found');
     });
 });
