+++ src/tests/validation/articles/articles.test.js
@@ -1,12 +1,12 @@
 const request = require('supertest');
-const app = require('../../../app');
-const User = require('../../../models/User');
-const Article = require('../../../models/Article');
+import app from '../../../index';
+import User from '../../../models/User';
+import Article from '../../../models/Article';
 const mongoose = require('mongoose');
 
 describe('Tests for validating the create articles feature', () => {
     process.env.API_BASE = '/api';
-    const userApiBase = process.env.API_BASE + '/users';
+    const userApiBase = process.env.API_BASE;
     const articleApiBase = process.env.API_BASE + '/articles';
     let access_token;
 
@@ -21,7 +21,7 @@ describe('Tests for validating the create articles feature', () => {
                 password: 'P@ssw0rd',
                 confirmPassword: 'P@ssw0rd',
             });
-        res = await request(app)
+        const res = await request(app)
             .post(userApiBase + '/login')
             .send({
                 email: 'nserekopaul@gmail.com',
